refactor: celery app moved to worker directory, it makes more sense there
